don't unlink output upon interupt if using zcat

+++ usr.bin/compress/compress.c
@@ -1,5 +1,5 @@
 #ifndef lint
-static char sccsid[] = "@(#)compress.c	5.3 (Berkeley) 06/16/85";
+static char sccsid[] = "@(#)compress.c	5.4 (Berkeley) 08/27/85";
 #endif not lint
 
 #define	min(a,b)	((a>b) ? b : a)
@@ -1326,7 +1326,8 @@ foreground()
 
 onintr ( )
 {
-    unlink ( ofname );
+    if (!zcat_flg)
+	unlink ( ofname );
     exit ( 1 );
 }
